requirements.txt
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-cors-headers
django-filter
djangorestframework==3.13.1
djangorestframework-simplejwt
drf-yasg==1.20.0
gunicorn==19.9.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
Pillow
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.8
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
